pyproject.toml
[project]
name = "anime-object-cut"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dghs-imgutils>=0.17.0",
    "fastapi[standard]>=0.115.12",
    "loguru>=0.7.3",
    "onnxruntime>=1.22.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.3",
]
